+++ src/main/java/it/gov/pagopa/wispconverter/controller/RedirectController.java
@@ -14,6 +14,7 @@ import it.gov.pagopa.wispconverter.exception.AppException;
 import it.gov.pagopa.wispconverter.service.ConverterService;
 import it.gov.pagopa.wispconverter.util.Constants;
 import it.gov.pagopa.wispconverter.util.ErrorUtil;
+import it.gov.pagopa.wispconverter.util.TraceReEvent;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
@@ -28,9 +29,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import java.time.Instant;
 import java.util.Map;
@@ -47,11 +46,14 @@ public class RedirectController {
     private final ErrorUtil errorUtil;
     private final MessageSource messageSource;
 
+    private static final String BO_REDIRECT = "redirect";
+
     @Operation(summary = "", description = "", security = {@SecurityRequirement(name = "ApiKey")}, tags = {"Redirect"})
     @ApiResponses(value = {
             @ApiResponse(responseCode = "302", description = "Redirect to Checkout service.", content = @Content(schema = @Schema()))
     })
     @GetMapping(value = "/redirect")
+    @TraceReEvent(businessProcess=BO_REDIRECT)
     public String redirect(@Parameter(description = "", example = "identificativoIntermediarioPA_sessionId")
                            @NotBlank(message = "{redirect.session-id.not-blank}")
                            @RequestParam("sessionId") String sessionId,
@@ -85,11 +87,13 @@ public class RedirectController {
 
     }
 
+
     @Operation(summary = "", description = "", security = {@SecurityRequirement(name = "ApiKey")}, tags = {"Redirect"})
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Redirect info to Checkout service.", content = @Content(schema = @Schema(implementation = RedirectResponse.class)))
     })
     @GetMapping(value = "/redirect", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @TraceReEvent(businessProcess=BO_REDIRECT)
     public ResponseEntity<RedirectResponse> redirectInfo(@Parameter(description = "", example = "identificativoIntermediarioPA_sessionId")
                                                          @NotBlank(message = "{redirect.session-id.not-blank}")
                                                          @RequestParam("sessionId") String sessionId) {
