+++ src/main/java/it/gov/pagopa/wispconverter/service/model/re/ComponenteEnum.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.wispconverter.service.model.re;
 
 public enum ComponenteEnum {
-    FESP,
+//    FESP,
+    WISP_SOAP_CONVERTER,
     WISP_CONVERTER;
 }
