+++ src/main/java/it/gov/pagopa/wispconverter/util/Constants.java
@@ -19,6 +19,8 @@ public class Constants {
     public static final String MDC_REQUEST_ID = "requestId";
     public static final String MDC_OPERATION_ID = "operationId";
 
+    public static final String MDC_BUSINESS_PROCESS = "businessProcess";
+
     public static final String MDC_CLIENT_OPERATION_ID = "clientOperationId";
     public static final String MDC_CLIENT_EXECUTION_TIME = "clientExecutionTime";
