+++ src/main/java/it/gov/pagopa/wispconverter/util/ReUtil.java
@@ -62,7 +62,8 @@ public class ReUtil {
                 .operationId(MDC.get(Constants.MDC_OPERATION_ID))
                 .clientOperationId(MDC.get(Constants.MDC_CLIENT_OPERATION_ID))
                 .componente(ComponenteEnum.WISP_CONVERTER)
-                .insertedTimestamp(Instant.ofEpochMilli(Long.parseLong(MDC.get(Constants.MDC_START_TIME))));
+                .insertedTimestamp(Instant.ofEpochMilli(Long.parseLong(MDC.get(Constants.MDC_START_TIME))))
+                .businessProcess(MDC.get(Constants.MDC_BUSINESS_PROCESS));
     }
 
     public static ReEventDto.ReEventDtoBuilder createBaseReInternal(){
