+++ src/main/java/it/gov/pagopa/wispconverter/util/TraceReEvent.java
@@ -1,0 +1,12 @@
+package it.gov.pagopa.wispconverter.util;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface TraceReEvent {
+    public String businessProcess() default "not-configured";
+}
