+++ src/main/resources/i18n/messages_it.properties
@@ -1,5 +1,5 @@
 error-code.ERROR.title=Errore del sistema
-error-code.ERROR.detail=Si è verificato un errore imprevisto. Si prega di contattare l'assistenza
+error-code.ERROR.detail=Si è verificato un errore imprevisto. Si prega di contattare l''assistenza
 
 error-code.GENERIC_ERROR.title=Errore generico flusso
 error-code.GENERIC_ERROR.detail=Si è verificato un errore durante il flusso di conversione. {0}
@@ -8,7 +8,7 @@ error-code.PARSING_GENERIC_ERROR.title=Errore generico di parsing
 error-code.PARSING_GENERIC_ERROR.detail=Si è verificato un errore durante il parsing del payload. {0}
 
 error-code.PARSING_INVALID_HEADER.title=Errore di parsing per SOAP Header
-error-code.PARSING_INVALID_HEADER.detail=Si è verificato un errore durante il parsing del payload. L'header SOAP nel payload non è corretto: {0}
+error-code.PARSING_INVALID_HEADER.detail=Si è verificato un errore durante il parsing del payload. L''header SOAP nel payload non è corretto: {0}
 
 error-code.PARSING_INVALID_BODY.title=Errore di parsing per SOAP Body
 error-code.PARSING_INVALID_BODY.detail=Si è verificato un errore durante il parsing del payload. Il body SOAP nel payload non è corretto: {0}
@@ -17,7 +17,7 @@ error-code.PARSING_INVALID_XML_NODES.title=Errore di parsing per XML
 error-code.PARSING_INVALID_XML_NODES.detail=Si è verificato un errore durante il parsing del payload. La lista dei nodi estratti dal documento XML deve contenere almeno un elemento.
 
 error-code.PARSING_INVALID_ZIPPED_PAYLOAD.title=Errore di estrazione ZIP
-error-code.PARSING_INVALID_ZIPPED_PAYLOAD.detail=Si è verificato un errore durante il parsing del payload. Impossibile effettuare l'unzip del payload.
+error-code.PARSING_INVALID_ZIPPED_PAYLOAD.detail=Si è verificato un errore durante il parsing del payload. Impossibile effettuare l''unzip del payload.
 
 error-code.PARSING_PRIMITIVE_NOT_VALID.title=Primitiva non valida
 error-code.PARSING_PRIMITIVE_NOT_VALID.detail=Si è verificato un errore durante il check della primitiva. La primitive [{0}] non è valida.
@@ -29,7 +29,7 @@ error-code.CONFIGURATION_INVALID_STATION.title=Stazione non trovata
 error-code.CONFIGURATION_INVALID_STATION.detail=Si è verificato un errore durante la generazione del carrello per il servizio Checkout. La stazione con codice [{0}] non esiste.
 
 error-code.PERSISTENCE_RPT_NOT_FOUND.title=RPT non trovata
-error-code.PERSISTENCE_RPT_NOT_FOUND.detail=Si è verificato un errore durante la ricerca dell'RPT. L'RPT con sessionId [{0}] not esiste.
+error-code.PERSISTENCE_RPT_NOT_FOUND.detail=Si è verificato un errore durante la ricerca dell''RPT. L''RPT con sessionId [{0}] not esiste.
 
 error-code.PERSISTENCE_REQUESTID_CACHING_ERROR.title=Errore di caching RequestID
 error-code.PERSISTENCE_REQUESTID_CACHING_ERROR.detail=Si è verificato un errore durante il caching della RequestID. {0}
@@ -53,7 +53,7 @@ error-code.CLIENT_CHECKOUT.title=Errore CheckoutClient
 error-code.CLIENT_CHECKOUT.detail=CheckoutClient - {0}
 
 error-code.CLIENT_CHECKOUT_NO_REDIRECT_LOCATION.title=Errore di redirect Checkout
-error-code.CLIENT_CHECKOUT_NO_REDIRECT_LOCATION.detail=Si è verificato un errore durante la comunicazione con il servizio Checkout. Nessun header 'Location' trovato.
+error-code.CLIENT_CHECKOUT_NO_REDIRECT_LOCATION.detail=Si è verificato un errore durante la comunicazione con il servizio Checkout. Nessun header ''Location'' trovato.
 
 error-code.CLIENT_CHECKOUT_INVALID_REDIRECT_LOCATION.title=Errore di redirect Checkout
-error-code.CLIENT_CHECKOUT_INVALID_REDIRECT_LOCATION.detail=Si è verificato un errore durante la comunicazione con il servizio Checkout. E' stato trovato un header 'Location' vuoto.
+error-code.CLIENT_CHECKOUT_INVALID_REDIRECT_LOCATION.detail=Si è verificato un errore durante la comunicazione con il servizio Checkout. E'' stato trovato un header ''Location'' vuoto.
